+++ packages/aws-lambda/src/index.js
@@ -1,7 +1,15 @@
 'use strict';
 
-const awsLambda = require('./aws_lambda');
+const { environment: environmentUtil } = require('@instana/serverless');
 
-module.exports = {
-  awsLambda
-};
+environmentUtil.validate();
+
+if (environmentUtil.isValid()) {
+  module.exports = exports = require('./wrapper');
+  // required for backwards compatibility for first four beta customers
+  exports.awsLambda = module.exports;
+} else {
+  module.exports = exports = require('./noop_wrapper');
+  // required for backwards compatibility for first four beta customers
+  exports.awsLambda = module.exports;
+}