+++ packages/aws-lambda/src/metrics/npmPackageName.js
@@ -1,11 +1,11 @@
 'use strict';
 
 const { util: coreUtil } = require('@instana/core');
+const { consoleLogger: logger } = require('@instana/serverless');
 
-const logger = require('../../util/console_logger');
 const rootDir = require('./rootDir');
 
-exports.payloadPrefix = 'npmPackageDescription';
+exports.payloadPrefix = 'npmPackageName';
 exports.currentPayload = undefined;
 
 exports.activate = function activate() {
@@ -15,7 +15,7 @@ exports.activate = function activate() {
     }
 
     if (!err && pckg) {
-      exports.currentPayload = pckg.description;
+      exports.currentPayload = pckg.name;
     }
   });
 };