+++ packages/aws-lambda/src/metrics/rootDir.js
@@ -1,0 +1,7 @@
+'use strict';
+
+const path = require('path');
+
+// We should be in /var/task/node_modules/@instana/aws-lambda/src/metrics initially, provide
+// /var/task as the starting dir for looking for the handler's package.json.
+exports.root = path.resolve(__dirname, '..', '..', '..', '..', '..');