+++ packages/aws-lambda/test/infra-mock/index.js
@@ -24,7 +24,7 @@ if (instanaKey == null) {
 
 /* dependencies */
 const https = sendUnencrypted ? require('http') : require('https');
-const constants = require('../../../src/util/constants');
+const constants = require('../../../serverless/src/constants');
 
 const legacySensorMode = process.env.LEGACY_SENSOR != null;
 const anotherLambda = process.env.ANOTHER != null;